Add common viame_utils module for shared utils
- Add type hint to run_pipeline param
