+++ server/viame_utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities that are common to both the viame server and tasks package."""
