+++ server/viame_utils/types/__init__.py
@@ -1,0 +1,1 @@
+from .pipelines import *  # noqa: F401 F403
